Refactor SkyWalking.Utilities
