+++ skywalking-csharp.sln
@@ -36,8 +36,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyWalking.Sample.Frontend"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyWalking.Diagnostics.HttpClient", "src\SkyWalking.Diagnostics.HttpClient\SkyWalking.Diagnostics.HttpClient.csproj", "{49DEFCA8-4289-4875-B6A5-35D84B3D2290}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyWalking.Extensions.DependencyInjection", "src\SkyWalking.Extensions.DependencyInjection\SkyWalking.Extensions.DependencyInjection.csproj", "{BF5579ED-113C-4EE6-AE03-9A9CA590C924}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyWalking.Diagnostics.CAP", "src\SkyWalking.Diagnostics.CAP\SkyWalking.Diagnostics.CAP.csproj", "{55621423-19C3-4928-8B45-666FA87BF6A2}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyWalking.Diagnostics.HttpClient.Tests", "test\SkyWalking.Diagnostics.HttpClient.Tests\SkyWalking.Diagnostics.HttpClient.Tests.csproj", "{05E533B1-473F-48C0-95F7-B87061BE9834}"
@@ -62,11 +60,17 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkyWalking.Sample.AspNet", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkyWalking.Transport.Grpc", "src\SkyWalking.Transport.Grpc\SkyWalking.Transport.Grpc.csproj", "{FF2FC3FB-D112-45C1-AF34-1FE37D55F682}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkyWalking.Extensions.Logging", "src\SkyWalking.Extensions.Logging\SkyWalking.Extensions.Logging.csproj", "{16F35181-4DB8-429D-9670-32A3DB7F0681}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkyWalking.Diagnostics.AspNetCore", "src\SkyWalking.Diagnostics.AspNetCore\SkyWalking.Diagnostics.AspNetCore.csproj", "{CFBB522D-495F-40F9-B68F-DA25E48D9F6A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkyWalking.Extensions.Configuration", "src\SkyWalking.Extensions.Configuration\SkyWalking.Extensions.Configuration.csproj", "{00D417E5-DC2D-4643-BFEA-DFD866C7A963}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Extensions", "Extensions", "{88AEFE60-F743-4AF3-8F55-B4B7B8079D95}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Utilities", "Utilities", "{4BD917BC-D579-4C75-9939-41BF012A4EE2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkyWalking.Utilities.Configuration", "src\SkyWalking.Utilities.Configuration\SkyWalking.Utilities.Configuration.csproj", "{0B87C4B8-3F4F-456A-B588-28D8E9E3395A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkyWalking.Utilities.DependencyInjection", "src\SkyWalking.Utilities.DependencyInjection\SkyWalking.Utilities.DependencyInjection.csproj", "{1812A19C-729C-410B-A90D-D5AF1404DD60}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkyWalking.Utilities.Logging", "src\SkyWalking.Utilities.Logging\SkyWalking.Utilities.Logging.csproj", "{AEB61F19-3DBB-411E-A67E-93007B96E160}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -106,10 +110,6 @@ Global
 		{49DEFCA8-4289-4875-B6A5-35D84B3D2290}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{49DEFCA8-4289-4875-B6A5-35D84B3D2290}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{49DEFCA8-4289-4875-B6A5-35D84B3D2290}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BF5579ED-113C-4EE6-AE03-9A9CA590C924}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BF5579ED-113C-4EE6-AE03-9A9CA590C924}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BF5579ED-113C-4EE6-AE03-9A9CA590C924}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BF5579ED-113C-4EE6-AE03-9A9CA590C924}.Release|Any CPU.Build.0 = Release|Any CPU
 		{55621423-19C3-4928-8B45-666FA87BF6A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{55621423-19C3-4928-8B45-666FA87BF6A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{55621423-19C3-4928-8B45-666FA87BF6A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -154,18 +154,22 @@ Global
 		{FF2FC3FB-D112-45C1-AF34-1FE37D55F682}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FF2FC3FB-D112-45C1-AF34-1FE37D55F682}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FF2FC3FB-D112-45C1-AF34-1FE37D55F682}.Release|Any CPU.Build.0 = Release|Any CPU
-		{16F35181-4DB8-429D-9670-32A3DB7F0681}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{16F35181-4DB8-429D-9670-32A3DB7F0681}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{16F35181-4DB8-429D-9670-32A3DB7F0681}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{16F35181-4DB8-429D-9670-32A3DB7F0681}.Release|Any CPU.Build.0 = Release|Any CPU
 		{CFBB522D-495F-40F9-B68F-DA25E48D9F6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CFBB522D-495F-40F9-B68F-DA25E48D9F6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CFBB522D-495F-40F9-B68F-DA25E48D9F6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CFBB522D-495F-40F9-B68F-DA25E48D9F6A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{00D417E5-DC2D-4643-BFEA-DFD866C7A963}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{00D417E5-DC2D-4643-BFEA-DFD866C7A963}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{00D417E5-DC2D-4643-BFEA-DFD866C7A963}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{00D417E5-DC2D-4643-BFEA-DFD866C7A963}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0B87C4B8-3F4F-456A-B588-28D8E9E3395A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0B87C4B8-3F4F-456A-B588-28D8E9E3395A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0B87C4B8-3F4F-456A-B588-28D8E9E3395A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0B87C4B8-3F4F-456A-B588-28D8E9E3395A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1812A19C-729C-410B-A90D-D5AF1404DD60}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1812A19C-729C-410B-A90D-D5AF1404DD60}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1812A19C-729C-410B-A90D-D5AF1404DD60}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1812A19C-729C-410B-A90D-D5AF1404DD60}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AEB61F19-3DBB-411E-A67E-93007B96E160}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AEB61F19-3DBB-411E-A67E-93007B96E160}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AEB61F19-3DBB-411E-A67E-93007B96E160}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AEB61F19-3DBB-411E-A67E-93007B96E160}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -181,7 +185,6 @@ Global
 		{80A67B09-83F2-477F-907F-95FFF5B8FEAF} = {844CEACD-4C85-4B15-9E2B-892B01FDA4BB}
 		{2B4E350E-A1E5-4B4A-A642-BCA6D3887E5A} = {844CEACD-4C85-4B15-9E2B-892B01FDA4BB}
 		{49DEFCA8-4289-4875-B6A5-35D84B3D2290} = {79ED86A5-E9B9-49B2-9354-C911C079D03E}
-		{BF5579ED-113C-4EE6-AE03-9A9CA590C924} = {79ED86A5-E9B9-49B2-9354-C911C079D03E}
 		{55621423-19C3-4928-8B45-666FA87BF6A2} = {79ED86A5-E9B9-49B2-9354-C911C079D03E}
 		{05E533B1-473F-48C0-95F7-B87061BE9834} = {613F0980-91ED-4064-8E8C-168582EF4AD7}
 		{4CF91A8C-25C8-4BD8-A93D-3183AF0FE2E2} = {613F0980-91ED-4064-8E8C-168582EF4AD7}
@@ -194,9 +197,12 @@ Global
 		{3A454A8A-A1B7-4C6D-BE08-52D956F95CC1} = {79ED86A5-E9B9-49B2-9354-C911C079D03E}
 		{C3649845-E458-448D-B97D-C7E231E1D040} = {844CEACD-4C85-4B15-9E2B-892B01FDA4BB}
 		{FF2FC3FB-D112-45C1-AF34-1FE37D55F682} = {CBFF7EE0-69D7-4D6A-9BBD-8E567FF4D810}
-		{16F35181-4DB8-429D-9670-32A3DB7F0681} = {79ED86A5-E9B9-49B2-9354-C911C079D03E}
 		{CFBB522D-495F-40F9-B68F-DA25E48D9F6A} = {79ED86A5-E9B9-49B2-9354-C911C079D03E}
-		{00D417E5-DC2D-4643-BFEA-DFD866C7A963} = {79ED86A5-E9B9-49B2-9354-C911C079D03E}
+		{88AEFE60-F743-4AF3-8F55-B4B7B8079D95} = {05BF0D4E-C824-4EC8-8330-36C1FC49910E}
+		{4BD917BC-D579-4C75-9939-41BF012A4EE2} = {05BF0D4E-C824-4EC8-8330-36C1FC49910E}
+		{0B87C4B8-3F4F-456A-B588-28D8E9E3395A} = {4BD917BC-D579-4C75-9939-41BF012A4EE2}
+		{1812A19C-729C-410B-A90D-D5AF1404DD60} = {4BD917BC-D579-4C75-9939-41BF012A4EE2}
+		{AEB61F19-3DBB-411E-A67E-93007B96E160} = {4BD917BC-D579-4C75-9939-41BF012A4EE2}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {94C0DA2C-CCCB-4314-93A2-9809B5DD0583}
