+++ src/SkyWalking.Diagnostics.AspNetCore/SkyWalking.Diagnostics.AspNetCore.csproj
@@ -17,7 +17,7 @@
     <ItemGroup>
         <ProjectReference Include="..\SkyWalking.Abstractions\SkyWalking.Abstractions.csproj" />
         <ProjectReference Include="..\SkyWalking.Core\SkyWalking.Core.csproj" />
-        <ProjectReference Include="..\SkyWalking.Extensions.DependencyInjection\SkyWalking.Extensions.DependencyInjection.csproj" />
+        <ProjectReference Include="..\SkyWalking.Utilities.DependencyInjection\SkyWalking.Utilities.DependencyInjection.csproj" />
     </ItemGroup>
 
     <ItemGroup>
