+++ src/SkyWalking.Diagnostics.EntityFrameworkCore/SkyWalkingBuilderExtensions.cs
@@ -18,7 +18,7 @@
 
 using System;
 using Microsoft.Extensions.DependencyInjection;
-using SkyWalking.Extensions.DependencyInjection;
+using SkyWalking.Utilities.DependencyInjection;
 
 namespace SkyWalking.Diagnostics.EntityFrameworkCore
 {
