+++ src/SkyWalking.Utilities.DependencyInjection/SkyWalking.Utilities.DependencyInjection.csproj
@@ -1,15 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <Import Project="..\..\build\common.props" />
   <PropertyGroup>
-    <Description>Microsoft.Extensions.DependencyInjection (IServiceCollection) support for SkyWalking.</Description>
-    <AssemblyTitle>SkyWalking.Extensions.DependencyInjection</AssemblyTitle>
+    <Description>Microsoft.Utilities.DependencyInjection (IServiceCollection) support for SkyWalking.</Description>
+    <AssemblyTitle>SkyWalking.Utilities.DependencyInjection</AssemblyTitle>
     <TargetFramework>netstandard2.0</TargetFramework>
-    <AssemblyName>SkyWalking.Extensions.DependencyInjection</AssemblyName>
-    <PackageId>SkyWalking.Extensions.DependencyInjection</PackageId>
+    <AssemblyName>SkyWalking.Utilities.DependencyInjection</AssemblyName>
+    <PackageId>SkyWalking.Utilities.DependencyInjection</PackageId>
     <PackageTags>SkyWalking;APM;Diagnostics</PackageTags>
     <PackageReleaseNotes>
     </PackageReleaseNotes>
-    <RootNamespace>SkyWalking.Extensions.DependencyInjection</RootNamespace>
+    <RootNamespace>SkyWalking.Utilities.DependencyInjection</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\SkyWalking.Core\SkyWalking.Core.csproj" />
